struct cmd

+++ Source/system/mylib.h
@@ -6,10 +6,10 @@
 #define DEBUG
 
 typedef unsigned char   uint8_t;
-typedef unsigned int    uint16_t;
+typedef unsigned short  uint16_t;
 typedef unsigned long   uint32_t;
 typedef char            int8_t;
-typedef int             int16_t;
+typedef short           int16_t;
 typedef long            int32_t;
 typedef uint8_t         u8;
 typedef uint16_t        u16;
